wall/wextra/werror

+++ src/CheckConfData.cpp
@@ -29,7 +29,7 @@ void Config::CheckClientMaxBodySize (std::string maxBodysize){
         txt = maxBodysize;
     else
         txt = maxBodysize.substr(0, maxBodysize.size() - 1);
-    for (int i = 0; i < txt.length(); i++)
+    for (size_t i = 0; i < txt.length(); i++)
     {
         if (!isdigit(txt[i]))
             throw std::invalid_argument("CheckClientMaxBodySize Error!");
@@ -39,7 +39,7 @@ void Config::CheckClientMaxBodySize (std::string maxBodysize){
 
 void Config::CheckLocation(std::vector<Location> location)
 {
-    for (int i = 0; location.size() > i ; i++)
+    for (size_t i = 0; location.size() > i ; i++)
     {
         CheckPath(location[i].path);
         CheckAllow(location[i].methods);
@@ -53,7 +53,7 @@ void Config::CheckPath(std::string path){
 }
 
 void Config::CheckAllow(std::vector<std::string> methods){
-    for (int i = 0; i < methods.size(); i++)
+    for (size_t i = 0; i < methods.size(); i++)
     {
         if (!(methods[i].compare("PUT") == 0 || methods[i].compare("GET") == 0 || methods[i].compare("POST") == 0 || methods[i].compare("DELETE") == 0))
             throw std::invalid_argument("Methods error");
